Use only one end-point for linking any respondent
