+++ src/main/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/service/CcdAccessService.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.reform.divorce.casemaintenanceservice.service;
 
 public interface CcdAccessService {
-    void linkRespondent(String authorisation, String caseId, String letterHolderId, boolean isCoRespondent);
+    void linkRespondent(String authorisation, String caseId, String letterHolderId);
 
     void unlinkRespondent(String authorisation, String caseId);
 }
