+++ src/test/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/functionaltest/LinkRespondentITest.java
@@ -61,6 +61,10 @@ public class LinkRespondentITest extends AuthIdamMockSupport {
         (String)ReflectionTestUtils.getField(CcdCaseProperties.class, "RESP_LETTER_HOLDER_ID_FIELD");
     private static final String RESP_RECEIVED_AOS_FIELD =
         (String)ReflectionTestUtils.getField(CcdCaseProperties.class, "RESP_RECEIVED_AOS_FIELD");
+    private static final String CO_RESP_LETTER_HOLDER_ID_FIELD =
+        (String)ReflectionTestUtils.getField(CcdCaseProperties.class, "CO_RESP_LETTER_HOLDER_ID_FIELD");
+    private static final String CO_RESP_RECEIVED_AOS_FIELD =
+        (String)ReflectionTestUtils.getField(CcdCaseProperties.class, "CO_RESP_RECEIVED_AOS_FIELD");
 
     private static final String API_URL =
         String.format("/casemaintenance/version/1/link-respondent/%s/%s", CASE_ID, LETTER_HOLDER_ID);
@@ -214,6 +218,34 @@ public class LinkRespondentITest extends AuthIdamMockSupport {
     }
 
     @Test
+    public void givenLetterHolderIdDoNotMatch_whenLinkCoRespondent_thenReturnNotFound() throws Exception {
+        final String message = getUserDetails();
+        final String serviceAuthToken = "serviceAuthToken";
+
+        final CaseDetails caseDetails = CaseDetails.builder()
+            .state(CaseState.AOS_AWAITING.getValue())
+            .data(Collections.singletonMap(CO_RESP_LETTER_HOLDER_ID_FIELD, "nonmatchingletterholderid"))
+            .build();
+
+        stubUserDetailsEndpoint(HttpStatus.OK, new EqualToPattern(USER_TOKEN), message);
+        stubCaseWorkerAuthentication(HttpStatus.OK);
+
+        when(serviceTokenGenerator.generate()).thenReturn(serviceAuthToken);
+        when(coreCaseDataApi.readForCaseWorker(
+            BEARER_CASE_WORKER_TOKEN,
+            serviceAuthToken,
+            CASE_WORKER_USER_ID,
+            jurisdictionId,
+            caseType,
+            CASE_ID)
+        ).thenReturn(caseDetails);
+
+        webClient.perform(MockMvcRequestBuilders.post(API_URL)
+            .header(HttpHeaders.AUTHORIZATION, USER_TOKEN))
+            .andExpect(status().isNotFound());
+    }
+
+    @Test
     public void givenNotAosAwaitingState_whenLinkRespondent_thenReturnNotFound() throws Exception {
         final String message = getUserDetails();
         final String serviceAuthToken = "serviceAuthToken";
@@ -325,6 +357,46 @@ public class LinkRespondentITest extends AuthIdamMockSupport {
         final CaseDetails caseDetails = CaseDetails.builder()
             .state(CaseState.AOS_AWAITING.getValue())
             .data(Collections.singletonMap(RESP_LETTER_HOLDER_ID_FIELD, LETTER_HOLDER_ID))
+            .build();
+
+        stubUserDetailsEndpoint(HttpStatus.OK, new EqualToPattern(USER_TOKEN), message);
+        stubCaseWorkerAuthentication(HttpStatus.OK);
+
+        when(serviceTokenGenerator.generate()).thenReturn(serviceAuthToken);
+        when(coreCaseDataApi.readForCaseWorker(
+            BEARER_CASE_WORKER_TOKEN,
+            serviceAuthToken,
+            CASE_WORKER_USER_ID,
+            jurisdictionId,
+            caseType,
+            CASE_ID)
+        ).thenReturn(caseDetails);
+
+        doNothing()
+            .when(caseAccessApi)
+            .grantAccessToCase(
+                eq(BEARER_CASE_WORKER_TOKEN),
+                eq(serviceAuthToken),
+                eq(CASE_WORKER_USER_ID),
+                eq(jurisdictionId),
+                eq(caseType),
+                eq(CASE_ID),
+                argThat(new UserIdMatcher(USER_ID))
+            );
+
+        webClient.perform(MockMvcRequestBuilders.post(API_URL)
+            .header(HttpHeaders.AUTHORIZATION, USER_TOKEN))
+            .andExpect(status().isOk());
+    }
+
+    @Test
+    public void givenAllGoesWell_whenLinkCoRespondent_thenProceedAsExpected() throws Exception {
+        final String message = getUserDetails();
+        final String serviceAuthToken = "serviceAuthToken";
+
+        final CaseDetails caseDetails = CaseDetails.builder()
+            .state(CaseState.AOS_AWAITING.getValue())
+            .data(Collections.singletonMap(CO_RESP_LETTER_HOLDER_ID_FIELD, LETTER_HOLDER_ID))
             .build();
 
         stubUserDetailsEndpoint(HttpStatus.OK, new EqualToPattern(USER_TOKEN), message);
